refactor: finalize methods and move classes to new package
